[config.py] Improve ConfigClock initialisation
There is a chicken-or egg initialisation issue in the ConfigClock `__init__()` code.  If this init code is called before the UsageConfig.py code is loaded then the config.usage dictionary items referenced are not yet defined.

I have moved the problematic initialisation code out of the ConfigClock.__init__() so that when ConfigClock is called during load time it won't need to access any of the config dictionary items.  I believe this is safe for the OpenViX and Beyonwiz builds and will require no other changes to resolve the current merge crash issue where ConfigClock needs to be called before UsageConfig can be loaded.

I have added a simple "if" in the handleKey() and genText() code that sees if the initialisation variables are initialised and if not then initialise them.  This processor intensive initialisation is then done only once during the first iteration of these functions that are frequently called.