Fix Bug #590: IceTV timer description issues
[IceTV.plugin]
Add code to check the EPG add eit ids and descriptions to timers
missing them, and update timer eit ids and descriptions if they
change in the EPG.

Some care is needed to account for the fact that in a description
contains a line end marker ("\r\n", "\n" or "\r"), which is preserved
when timers.xml is written, but is converted to a single space when
timers.xml is read.

Using the fix for Bug #591: Lookup errors in eEPGCache::lookupEventTime()
should reduce the likelihood of having missing eit ids and descriptions
in timers when they are created, but it is not necessary for this
fix to work.

[IceTV.API]

Update IceTV plugin version.