Fix Bug #592: Crash on zap in RecordTimer when timeshifted
Removes the call to self.InfoBarInstance.timeshiftActivated()
and always inserts "Save timeshift and zap" in the first choice slot
if self.InfoBarInstance.save_timeshift_file is False (since
openChoiceActionBeforeZap() is only called when timeshift is being viewed
the test is redundant anyway).

Removes the "Save timeshift in movie dir and zap"/"save_movie" menu choice
and its associated code, since it does nothing in OpenViX or easy-ui-4.

Replaces the code in the "save" choice in zapAction() to use
self.InfoBarInstance.SaveTimeshift() instead of
self.InfoBarInstance.saveTimeshiftFiles().

Resets InfoBarTimeshift.save_timeshift_file in
self.InfoBarInstance.SaveTimeshift().

It's not entirely clear to me what the role of
InfoBarTimeshift.save_timeshift_file is intended to be, and I suspect
that it may not achieve what is intended.