[BackupRestore] Place backups in a Beyonwiz specific, versioned directory
This is solving two problems:

1. Backup location clashing between distros (eg. OpenViX and Beyonwiz)
2. Some backups can cause issues if / when applied to a firmware version
   that is not compatible with the stored values. By having versioned
   directories, we can avoid issues just by incrementing the directory
   version number. This is a lot more reliable than instructing users to
   not restore settings.

A benefit of the versioned approach is that it should be easier to write
scripts that can migrate settings backups from one version to another, as
we should have some idea as to what the incompatibilities are from version
to version. After all, the version bump would have been prompted by a
specific incompatibility encountered during development and testing.