Fix Bug #579: Recording start/end marks set to last "now" event rather than the last
The fix changes the setting of m_event_timestamps from using
map::operation[]() to using map::insert(), so that the first value
set for the event stays in the map if there is an attempt to update
it. The debug output has been changed to indicate cases where there
has been an attempt to change an existing m_event_timestamps entry
for an event.