[dvb] Increase NIT timeout to 30 seconds
Some Australian broadcasters fail to send NIT every 10 seconds as required
by the standard. This can result in services being found during a service
scan, but if the NIT is not available (due to timeout) the LCNs will not
be parsed and will be missing. As a result the scanned services will not
appear in the LCN database or the Terrestrial LCN bouquet.