Fix "Auto" recording timers sometimes not shutting down (#182)
* [Navigation] Fix debug message

In Navigation.recordService(), fix the "recording service"
debug message so that it prints the serviceref correctly
whether it's a ServiceReference or an eServiceReference.

* [eDVBLocalTimeHandler] Symbolic name for "time is ok" test value

Add eDVBLocalTimeHandler::timeOK for the time 00:00 1/1/2004 UTC used to
test whether system time has been set.

Also correct the value so that it is actually 1/1/2004 and not
24/12/2003 (the old value is (2004-1970) * 365 * 24 * 60 * 60,
which doesn't account for leap years).

* [Navigation] Fix "Auto" recording timers sometimes not shutting down

On a PVRs where the RTC resets on shutdown and where "auto" timers
are supported: if there's no Internet available at startup, at the
time when it's being checked whether the startup is for a recording
timer with "auto" set for its "After event" action, the correct
system time is not set, and so the test for that condition fails.

That then means that the system will not go to standby to complete
the recording, and at the end of the recording, the system will not
shut down.

If there is a valid broadcast service set for startup, the time
will be set about 10 sec after the test for the "auto" timer.

The test that fails when it should succeed is in Navigation.__init__():

        if getFPWasTimerWakeup():
            self.__wasTimerWakeup = True
            if nextRecordTimerAfterEventActionAuto and abs(self.RecordTimer.getNextRecordingTime() - time()) <= 360:  # <<< This test fails because the system time has not been set yet
                print '[Navigation] RECTIMER: wakeup to standby detected.'

The problem occurs whether MENU>Setup>System>Time settings>Sync
time is set to "TV broadcast" or "NTP (Internet)".

The T2 is not affected by this problem because it's not capable of
distinguishing a timer startup from manual startup.

Workaround is to ensure that an Internet connection is always
available at startup.

Replication steps

Do this this with MENU>Setup>System>Time settings>Sync time using
set to both "TV broacast" and "NTP (Internet)".

Set a recording timer in the near future with its "After event"
action set to "auto".

Leave a TV on and connected to observe system behaviour.

Shutdown the system, disconnect its Ethernet cable and wait for the
timer to start.

The system starts normally, and goes to fully running. However, it
does not go to standby 15 sec after the above test as intended.

When the recording finishes, the system does not return to shutdown,
but continues in live TV.

Fix:

Detect that the system time has not been correctly set when
Navigation.__init__() is called, and if the time has not been set
correctly, arrange for the code that should be run on a timer wakeup
to be re-run after the system clock has been set.