If at startup config.tv.lastservice is not valid or, if it is valid, but not a DVB service and the time is < 1 Jan 2004, find the first valid DVB service in the current bouquet, or if that fails and config.usage.multibouquet is set, then the first valid DVB service in any other bouquet.
As well as fixing this bug, the fix will try to find a valid DVB
service if config.tv.lastservice is not valid.

The new methods ChannelSelection::findServiceBouquet() and
ChannelSelection::forceZap() are intended to be able to be called
to replace most of the code in RecordTimer::failureCB(). They may also
be able to be adapted to replace most of the code involved in
zapping by channel number (and the channel panic button action).