Basic U4 remote updates
New key codes added to keyids.py and lib/python/Tools/KeyBindings.py.
Allow for known duplicate mappings of code -> code name when inverting
KEYIDS: KEY_VMODE and KEY_SWITCHVIDEOMODE both map to 227.

KEY_VMODE changed to KEY_SWITCHVIDEOMODE in ../../keyids.py,
lib/python/Tools/KeyBindings.py and keymap.xml.

Minor modification to lib/python/Components/Sources/HelpMenuList.py
to exclude help entries that are in the remote's
Tools.KeyBindings.keyDescriptions, but not in the remote's
rcpositions.xml file (e.g the entries for the RADIO, PIP and WEB
on the U4 remote and the entries for WIZTV (HDMI-IN), TV-OUT,
AUTOTIMER, SLOW, ABOUT and SETUP on the T series remotes).

This enables the correct actions on WIZTV (HDMI-IN) and SETUP.
TV-OUT is active, but only changes aspect ratio, not resolution as
intended.