Timer repeat code cleanup and simplification
In TimerEntry.nextRepeatTime(), change the sense of the day array
so that it is the same sense as the repeated flags, rather than
their inverse.

Simplify the test in the main loop so that it more obviously
corresponds to the comment that describes it, and remove an extraneous
test on the day repeat flag.

Change the comment to reflect the actual (unchanged) tests on
localbegin, localend and localnow.

Move the mktime(localrepeatedbegindate) out of the loop, since it's
constant over the loop.

In TimerEntry.nextRepeatTime(), use a simple list comprehension to
generate the day array from the repeated flags instead of a loop
to build the list by appending, and change the sense of the day
array so that it is the same sense as the repeated flags, rather
than their inverse.

This is a preliminary step to implement issue #23: Fill in timer
name from EPG.