Fix Issue #23: Fill in timer name from EPG
Also set a sensible event id in the timer from the EPG.

Setting an appropriate event is a major motivation for this change,
to support changes made in the IceTV plugin to fix Bug #603: Gaps
in IceTV EPG.  This fix in the event id will also mean that if the
timer times are changed to a different event, the Change Timer menu
should be available on (one of) the events highlighted as part of
the recording.

In TimerEntry, add methods getEventInfo(), updateEventInfo(),
nameInfoUpdated() and descriptionInfoUpdated().

Call them appropriately from notifiers on config items tha affect
when the timer runs, or when the name or description is changed by
the user.

User-entered description and name entries over-ride automatic
updating from the EPG.