libopen: also override accept().
This fixes the issue on VU+ Ultimo4k where descrambling stops
after enigma is set to standby. It's caused by an accepted socket
fd that wasn't closed when exec()ing bsa_server. Override accept()
as well, as to set the fd to FD_CLOEXEC as well, avoid leaking it
to children, as we already do for sockets created using open() and
socket().

NOTE: this code is ugly, it is copy&pasted from the other ugly
code. This really needs cleanup. Even better, try to get rid of
the wrappers altogether.