Enhance number skips
The "Enable skip for number buttons" option has switched from boolean
to multiple choice, allowing the numbers to skip in media files, but
still zap in timeshift, or to disable zap altogether.

Provide a new option "Number skip method" to choose between the current
skipping by relative seconds or a new "absolute percentage" method,
which will skip to a multiple of 10% from the start (e.g. press 5 to
jump to 50%).