FileTransfer: Only copy extra .ts files
Using the file transfer task (as used by the File Commander) would
effectively replace the extension with `*`; if no extension was present
then `.*` would be added, preventing the file from being transferred at
all.  Presumably this is intended to copy the additional files
associated with movies, so only replace `.ts` with `.*`.