Fix changing channel after Single EPG change
Changing channel from the Single (enhanced) EPG will remember the
channel on exit.  This was good when I was browsing the guide, saw
something in the background I wanted to watch, so exited the guide,
watched it, then restored the guide, continuing where I left off.
However, it was not good when using CH+, CH- or FAV, when I expected to
move from the channel I was on.  Explicitly set the current selection to
the current service before using those keys.