Allow multiple selections in the movie list
Use `0` to enter selection mode and select the current item.  Then use
`0` to toggle the selection, or `OK` to toggle and repeat the previous
`UP` or `DOWN`; `EXIT` will end selection mode.  Changing directory is
disallowed, as is rename.  Plugins will be passed the first item, along
with the selections in `serviceList`.  Move and delete will end the
selection (if all items were deleted).  `PLAY` or `PLAYPAUSE` will play
items in selected order, auto-resuming with no notifications.

The icons used for selection are skinnable as `icons/sel_off.png` and
`icons/sel_on.png`.