Fix subtitle synchronisation bugs
U4 hardware can have a very large gap between current PTS and the PTS
in the teletext subtitle pages that come from subsink.

There are also no guarantees that subtitle pages will arrive in
chronological order, so ensure that the pending list is sorted by PTS.

Also ditch the subtitle_noPTSrecordingdelay option, since the results
of using it are out of order subtitles.