Recognise TIMER/MEDIA/EPG from each other
Allow TIMER, MEDIA and EPG from all their screens (but not from the main
menu).  The current screen will EXIT and then open the new (which in the
case of the EPG will be the default, not the EPG originally used).

Pressing TIMER from EPG would exit back to the EPG, but to be consistent
with the other screens it now exits back to live TV (long GREEN can be
used to open the timer and exit back to the EPG).  The single channel
EPG (from the graphical EPG) is the exception: disable MEDIA and leave
TIMER exiting back to the EPG.