Reset File Commander's "temporary" directories
When the File Commander brings up its menu, or switches to selection
mode, it saves the current directories to temporaries, but only if the
current directories aren't the mount points.  In that case, the previous
temporary directory was used.  Set the temporaries to blank when in the
mount points (and explicitly prevent selection mode).