[skin.py] Add objectTypes skin attribute
This new attribute was created by OpenPLi and OpenViX to address issues with overlapping text in the COLOUR buttons when the Label() and Button() objects are overlapped with the StaticText() object.

The syntax of the objectTypes attribute is objectTypes="screen_element,object_type[,object_type,...]".  The use of an objectType attribute fully replaces the use of the simple "conditional=" attribute.

This change when adopted into the Beyonwiz skins will allow the Button() and Label() object code changes made by Prl to be reverted to the Enigma2 standard.

Adding this change will have no effect on the Beyonwiz build until/unless skins are coded to use the new attribute.  OverlayHD is adopting these changes now to support its use on both OpenViX and OpenPLi.