Commit: a155b11d3fd7530a0b4060a8bc219fa4ffbf0035 [a155b11] Parents: c5048ae941 Autor: Athanasios Oikonomou <athoik@gmail.com> Datum: Dienstag, 13. Februar 2018 21:04:14 Committer: Athanasios Oikonomou Use 32bit integers on eActionMap.bindAction
When using enigma2 on 64bit machine the eActionMap.bindAction fails with OverflowError.

That happens because the sys.maxint value on 64bit is 2**63-1 instead of 32bit value 2**31-1.

Define the maxint value instead of using the sys.maxint value.

This commit fixes the following error:

  File "/usr/local/lib/enigma2/python/Screens/InfoBar.py", line 66, in __init__
    x.__init__(self)
  File "/usr/local/lib/enigma2/python/Screens/InfoBarGenerics.py", line 136, in __init__
    eActionMap.getInstance().bindAction('', -maxint -1, self.actionA) #highest prio
OverflowError: in method 'eActionMap_bindAction', argument 3 of type 'int'