Add more keys, tweak help
Use REW and FF to toggle the current item (entering selection mode if
necessary) and move up or down, respectively.  Use long 0 to invert the
selection (entering selection mode if necessary).  Only invert items of
the same type (directory or file) as the current; the trash will be
deselected, but not selected.

Move 0's help into "Selection and exit" and add a help string for the
other numbers.