Fix bug #627: Incorrect ratings icons sometimes displayed
Clear the picon displayed in Components.Renderer.Pixmap when changed()
receives a CHANGED_CLEAR.

CHANGED_CLEAR is issued upstream in sources like Event and EventInfo
when no event can be found, and so this change clears the picon
when no ratings information can be found because no event has been
found.

Previously CHANGED_CLEAR was ignored.