gpixmap: introduce blit acceleration threshold
Until now, we only had a threshold for the entire surface to be
accelerated.
It is possible that we only blit a small portion of the surface, in wich
case it might be faster to use unaccelerated blit, even if the surface
itself is accelerated.

For now, use a default of 0, so the behaviour remains the same:
accelerate all blits on accelerated surfaces, no matter how small the
size of the blit area is.