frontend / sec: handle FE_TIMEDOUT while turning motorized systems
With a motorized rotary satellite dish setup, when power measurements
are disabled, a single tune request is issued when the motor starts to
rotate.
Every 250ms, the frontend lock status is checked, and the dish is
assumed to have reached the desired orbital position when it has had a
lock for 5 consecutive times.

However, this only works assuming that frontend drivers (or hardware)
will keep trying to get a lock forever, without using any internal
timeouts.
This is not the case. Most frontend drivers will eventually timeout,
reporting an FE_TIMEDOUT event.

In that case, we would have to wait for the hardcoded overall motor
timeout (of a very conservative two seconds per degree) before we would
attempt to tune again.

Instead, check for and handle FE_TIMEDOUT events, and re-issue the tune
request when a timeout occurs.