[elistboxcontent] Fix ConfigText alignment issues
For example, in a ConfigText field when you edit a text string the text relocates to the top of the text box rather than remaining centred. This commit fixes that and similar alignment issues for ConfigText and all sub classes of it, e.g. ConfigPassword, ConfigNumber,etc.

Original code: IanSav, Adapted for OpenPLi: Huevos.