Only use blit hardware acceleration for cases where the hardware can do so
It is not possible to implement blitAlphaTest easily using hardware
acceleration because the hardware does not have a facility to perform
comparisons on pixel values as part of the blit process. Use hardware
accelerated blits only in the copy and blend modes, perform alpha test
in software.

Partial fix for issue #629