Prevent initial playback subtitle seek
When a recording is played back with subtitles, the subtitle seek to
flush the buffers seems to be using the PTS of what was last played, not
the begin PTS of the recording (this may only affect the U4).  Since a
flush is unnecessary when first played, prevent that seek.