Fix bug #637: [eStaticServiceDVBPVRInformation] Extraction of name & time from filename can fail
Remove static function looksLikeRecording() which is both not
specific enough and fails to detect "short" and "event first"
recording filenames.

In eStaticServiceDVBPVRInformation::getName() replace it with inline
code to more accurately detect recording filenames and extract the
recording name and recording time from the filename.

For "short" format names, which only have the recording date in
their names, use the fallback recording creation time rather than
the date in the filename if the fallback time is not 0.

Set stm.tm_isdst to -1 before calling mktime() so that it uses
the zoneinfo database to determine whether to apply DST to the
time conversion.

In eStaticServiceDVBPVRInformation::getInfo() and
eStaticServiceDVBPVRInformation::getInfoString() call getName() to
ensure that any fields that are needed to be extracted from the
filename are extracted.