Fix crashes due to execute
Executing cuts would stop the service, thus when closing the editor
there would be no service for deleting the resume point.  Store the
service in self, rather than retrieving the currently playing service.

If execute ran very quickly its message would cause a modal error.	That
crash is caught in `MovieCut`, but a side effect is that cancelling the
execute doesn't close the editor.