Fix issue #649: Hide Enable EIT EPG when IceTV is enabled
Only show setting MENU>Setup>TV>EPG settings>Enable EIT EPG if IceTV
is disabled (or not installed).

Also prevent crashes in Screens/Setup.py if a non-existent config
variable name is used in a setup.xml "requires" attribute (e.g. if
it refers to a plugin setup variable).  Print an error message to
the debug log instead.

I originally thought this would be necessary to implement the test
on config.plugins.icetv.enable_epg, but it is created even if the
IceTV plugin is not installed.