Fix realignment of trailing spaces
Fix a trailing space bug introduced in commit ccd9a56 - trailing
spaces need to be moved when realigning for dirCenterIfFits, dirRight,
dirCenter and dirBidi alignment types.

This was the cause of odd highlighting issues in the MinuteInput
popup box as reported in the forum, because un-moved trailing spaces
were highlighted.

There were also similar issues with not moving trailing spaces for
dirBlock.