Fix issue #656: Refactor Input & InputBox in FileCommander
Remove module Plugins.Extensions.FileCommander.Input and re-base
Plugins.Extensions.FileCommander.InputBoxmod.InputBox on Screens.InputBox.

The only functionality required in the new rebased InputBoxmod.InputBox
is to add an ActionMap to it that supports the use of PLAY/PAUSE
to toggle insert/overwrite.

The previous mappings of KEY_REWIND and KEY_FASTFORWARD to seekBack
and seekForward and to the HOME and END actions on "break" respectively
in the InfobarSeekActions context are made dummy actions, because
the actions are already covered by mappings of the buttons to "home"
and "end" on "make" in the KeyboardInputActions context in the
ActionMap in Screens.InputBox.

The only functionality that InputBoxmod.InputBoxmod added is to
provide a default skin, so it simply inherits from InputBoxmod.InputBox
and adds the default skin.

InputBoxmod.PinInput is unused in FileCommander, so it has been
removed.  The old version only added similar functionality to that
added by InputBoxmod.InputBox, and that extra functionality is
probably unnecessary.