Fix issue #657: Refactor FileList in FileCommander - getParentDir() fixes
Implement the setting of self.parent_directory in
FileCommander.MultiFileSelectList so that
MultiFileSelectList.getParentDir() works correctly.

Always set self.parent_directory to None when it is invalid (not
sometimes None and sometimes False).

Change FileCommanderScreen to test for the new invalid value.

This is necessary to test the next step, and to allow the implementation
of navigation in MultiFileSelectList if that is wanted at a later
time.

This is not possible to test without modifying FileCommanderScreen,
but it has been tested that way.