Fix issue #657: Refactor FileList in FileCommander - FileListmod logic cleanup
Clean up some of the logic in FileListmod:

Replace if statements with the use of Boolean variables or test
results and changing some explicit looping tests to x in list tests.

Fix the logic of alreadySelected in the if self.showFiles part
of MultiFileSelectList.changeDir() so that full pathnames are
compared instead of basenames.

Reorder & simplfy isDir/isLink tests in FileEntryComponent() and
MultiFileSelectEntryComponent().

This is not possible to test the changes in MultiFileSelectList.changeDir()
that test for alreadySelected without modifying FileCommanderScreen,
but it has been tested that way.