Fix bug #664: MovieSelection sometimes sets all configurable buttons to add AutoTimer
In MovieSelection.updateUserDefinedActions() change the call of
btn.setChoices() to include default=btn.default so that the selection
retains the correct default.

Also, initialise and update the configurable button ConfigSelections
with copies of the selection list to avoid potential future issues
with aliasing between the selection lists in different ConfigSelections.