Fix bug #669: Make FileCommander setup screen a more standard setup
Refactor FileCommander.FileCommanderConfigScreen to inherit from
Screens.Setup and load its setup definitions from FileCommander/setup.xml.

Add setup.xml with setup definitions.

Change the definition of config.plugins.filecommander.path_default
from ConfigText to ConfigDirectory so that it doesn't respond to
text input, displays an empty string as "List of storage devices",
and uses that as the default path.

Fix bug in goDefaultfolder that when
config.plugins.filecommander.path_default is not set, it doesn't
go to "List of storage devices" when 5 (Go to default folder) is
pressed.

Change the definition of goRestart() to allow for the fact that
Setup can call close() without arguments.

When setting Default folder, use LocationBox instead of MovieLocationBox,
because MovieLocationBox restricts the directories that can be set
for Default folder, and the FileCommander isn't restricted from
examining any directories.