Fix issue #672: Add interfaces to useful file information programs to FileCommander
Add actions on remote buttons 6, 7 and 9 to run "file" and "ffprobe"
on buttons 6 & 7, and a user-configurable set of checksum calculations
(defaulting to MD5) on button 9.

There is code to allow button 8 to run "mediainfo", but access to it
is commented out because it's currently not available as a Beyonwiz
package.

Add an entry to setup.xml for the hash set selection config.