Fix bug #660: Changes to FileCommander extension filter doesn't update when changed
Create methods in key_actions to return the currently active extension
filter regular expression and to return a tuple containing the
extension filter settings.

Use the new fileFilter() method to set the initial filter settings
and to update the settings in the file lists when the settings
change.

Use the new filterSettings() method to compare the filter settings
before and after the settings page is opened, and update the filters
in the file lists when they change.

Remove unnecessary use of settings
config.plugins.filecommander.path_left_tmp and
config.plugins.filecommander.path_right_tmp in goMenu() and goReset().

Removed unused goMenu() method in FileCommanderScreenFileSelect.