Fix bug #663: File commander sometimes silently fails to execute script
In key_actions:
Change do_run_script() so that it checks that the script is readable
and issues an error MessageBox if not.  Checks whether the script
is executable and runs it directly if it is and runs it as "sh
script" if it is not.

In  do_run_script(), clean up the use of self.commando where the
length-one tuple is used to create a new length-one tuple:
(self.commando[0],).  Change those cases to simply use self.commando.