Fix issue #661: File commander extension filters incomplete
Construct the File Commander media type filters from
Plugins.Extensions.FileCommander.FileList.EXTENSIONS
and from Plugins.Extensions.FileCommander.MOVIEEXTENSIONS
so that the filters agree with the file type dicts.

Make the filters more specific by anchoring the regular expressions
at the string end, and bu explicityl enumerating the correct extensions
in the records regular expression.

Add a call of onLayout() to update the screen title to show/hide
the "(*)" "filters enabled" flag when the filter settings change.

In Components.FileList, add "jpe" as a "picture" extension and "iso"
as a "movie" extension to match the old regular expressions in
Plugins.Extensions.FileCommander.FileList and continue
to have them in the new extensions.