Fix issue #679: Add .log as a text file type in File Commander
Add the ".conf" & ".srt" extensions that were in
FileCommander.FileList.LOCAL_EXTENSIONS but not in
addons.key_actions.TEXT_EXTENSIONS to TEXT_EXTENSIONS (and make
them editable).

Remove all the mappings to "txt" in LOCAL_EXTENSIONS, and construct
then from TEXT_EXTENSIONS, but if the extension is already in
LOCAL_EXTENSIONS, it isn't over-ridden by the update. E.g. although
".py" is in TEXT_EXTENSIONS, that doesn't change the "py": "py"
mapping in LOCAL_EXTENSIONS.

This has the side-effect of adding ".log" (as well as some other
extensions) as a text file type for displaying the file type icon.