Fix bug #684: Crash in File Commander when file used (OK) directly after move
Test for the existence of files when they are directly used in
system calls, and display a popup if the file is in the
FileList but doesn't exist in the file system.

Also clean up code and error handling in addons.key_actions.onFileAction()
in the fallback case where Scanner is used to to try to find a
handler for the file type - in guess_type() pass the full file path
and not just the filename, and use the already-calculated file path
to pass to the scanner instead of re-creating it. In the error
handling display the error popup if openFile() returns False as
well as when there's a TypeError in openFile.