Fix issue #673: Tidy up FileCommander list headers
Add new method key_actions.statInfo() to return a tuple of formatted
stat data for use in the FileCommander file list display headings.

In FileCommanderScreen and FileCommanderScreenFileSelect, replace
the single file list header Label containing both the file path
name and the file's stat information with a Label to hold the path
name and a List/TemplatedMultiContent/Listbox render chain to display
data from key_actions.statInfo(). This allows the data to displayed
to be selected and placed by the skin author.

Update the FileCommanderScreen and FileCommanderScreenFileSelect
fallback skins accordingly.

Simplify the structure of updateHead() by ignoring the values of
self.SOURCELIST, self.TARGETLIST and updating the left and right
headers in a loop, and update from key_actions.statInfo()
instead of from key_actions.Info().

Replace explicit setting of the header text with calls to updateHead().

Change the import from addons.key_actions to use explicit imports
instead of *.