[FileCommander] Fix problems with import of Plugins.Extensions.PicturePlayer
FileCommander crashes if Plugins.Extensions.PicturePlayer is not
installed.

Also, the import statement

from Plugins.Extensions.PicturePlayer.ui import config

is a little confusing, because the purpose of the statement is to
import Plugins.Extensions.PicturePlayer.ui and initialise the
config variables in it, not to import its config variables
for use in the importing module.

Remove the import of PicturePlayer in FileCommander.plugin
and replace it with code in FileCommander.addons.type_utils
that checks whether Plugins.Extensions.PicturePlayer.ui is installed.

If it isn't installed, pop up an error message instead of crashing.