[FileCommander] Refresh automatically after jobs
Move and copy operations are done via jobs and removing a directory is
done via a console container.  This generally means that a manual
refresh is necessary to update the list.  Use a callback on completion
to refresh automatically.  Update the title with the number of active
jobs (started by the current FC, not in general; it's also not present
during multiselection).  Resolves #686.