Fix Bug #695: File Commander doesn't detect failure to unpack because of missing archive program
Fix bidirpipe() in console.cpp so that in the child fork it returns
a non-zero exit status to indicate an error if the execvp() call
fails. 127 was chosen as the exit status to match what system(3)
returns in similar circumstances. It's also similar to what bash(1)
does when sh -c can't find a program.

Also have bidirpipe() print an appropriate error message to the
child's stderr.  The old message was incorrect.

Modify eConsoleAppContainer::readyRead() so that it passes -1 as
the argument to the appClosed() PSignal if the child process was
terminated by a signal, and ignores the return from waitpid() if
the process wasn't exited or terminated, e.g. if it received a stop
or continue signal.