Fix Bug #710: [IceTV] Unnecessary updates on unchanged timers
[plugin]
In EPGFetcher.processTimers(), instead of calling onTimerChanged(timer)
when an IceTV timer update leaves the timer unchanged, simply mark
the timer as in "pending" state, give it a message that indicates
that the timer was already up to date, and queue the response.

This prevents unnecessary delete/re-create operations on IceTV timer
when an IceTV update on a timer doesn't change the timer.

[API]
Update the version datestamp.