Refactor downmix detection and setting
To fix bug #714, a mechanism to write "passthrough" into
/proc/stb/audio/dts is required.

A clean way to do that is to move the code in setDTSDownmix() into
a new local function and allow that to be called with a bool argument.

That can be extended to all three setXXXDownmix() functions, and
the mostly copy/paste code to detect downmix capability can also
be put into a single function.

The two new functions hasDownmix() and setDownMix() take an argument
to specify the audio format to be tested.