Fix bug #715: Zap timer startup sometimes blocked when user is in MovieSelection
In RecordTimer.activate() only try to close the MoviePlayer instance
on zap if the MoviePlayer can be closed immediately (i.e, is in
execing state).

Also, when MoviePlayer is closed, don't increment the timer's start
time - it's not necessary, because activate() will be re-called
after the close() has done its deferred processing even if the start
time isn't incremented.