InfoBarShowHide: Stop hide timer on hide
The timer used to automatically hide the infobar is not stopped when the
infobar is hidden by pressing INFO, so the info would disappear.  Stop
the timer in the on-hide event.  (This supercedes a previous patch that
only stopped the timer on EXIT, which didn't work in the movie player.)