Single EPG: Call it that
The single EPG is labelled differently to the other EPGs, having "Show"
in front of it.  This causes the help message to duplicate "Show", and
is just inconsistent in general.  Label it as "Single EPG", to
complement "Multi EPG", and change the help text to "single channel"
(without a hyphen).  (Unfortunately, it does mean resetting it as
default and/or restoring its position in the menu.)