Screens.Console: Write output to file
In addition to displaying Console output on the screen, write each
command, its start time and output to `/home/root/logs/console.txt`;
end with the finish time.  Once finished, GREEN can be used to view
the file ("details" of the execution).  The file is overwritten with
each new Console (currently needs to be preserved manually).