keymap: Incorporate/remove devices
The T-series remote control is called "dreambox advanced remote control
(native)"; the U4 is "dreambox remote control (native)"; and the V2 is
"dreambox advanced remote control".  The keymap uses the first name to
selectively define some keys (for the DVD player in particular).  Rather
than adding new devices for the U4 and V2, incorporate the keys directly
(but not the SONG ones, as the Beyonwiz remotes lacks them) and remove
all the other devices (except for "keyboard").  (The T-series does have
"dreambox front panel", but the mapping is not used; the U4 does not
appear to define a name; the V2 is "Hisi Front Panel Button Controls").