Fix bug #727: [IniLCNScanner] Fails to propagate some changes to bouquets.* files
Change addInBouquets() to only short-circuit the update if the new
entry for the Terrestrial LCN bouquet file is identical with a
single entry in the file. If the entry is to be updated, do it in
the position of the first occurrence of any entry for the Terrestrial
LCN bouquet file. If there is no existing entry, put the Terrestrial
LCN bouquet file entry before all other existing service entries
in the file.

Done with a complete re-write of addInBouquets() to clean it up a
bit and make it more Pythonic.