InfoBarCueSheetSupport: Adjust resume without length
Resume always asks if the length of the movie is unknown.  This is not
needed if the movie always starts from the beginning.  It would also
default to "no" if the movie should always resume.  I'm tempted to
remove the question altogether in that case, but settled on defaulting
to "yes".