Add option to hide visible bouquet markers.
This allows visible markers (bouquet spacers i.e. #SERVICE 1:320:0:0:0:0:0:0:0:0:) such as those used by fastscan and cable scan plugins to be hidden as a user defined option. When in "channelSelection" move mode the markers will be displayed allowing channels to be placed between them if desired.

It looks like OpenATV have had all the CPP code to do this installed already for years but no Python code to control it.