Add IceTV "Violence" genre
This changes the IceTV remappings of genre names with the same genre
eit value where the remapped eit value is 0x00-0x0f.

That means that the IceTV EPG must be reloaded for the changes to
take effect, but that reload will happen after the normal reboot
on completion of the update.

This fix is related to, but not the cause of, IceTV plugin crashes
that happened on 2019-09-20.

Bump IceTV plugin version number.