Icetv batch fetch (#1422)
* Fetch only user's selected channels to the IceTV channel map

Add UserChannels class to the IceTV API. UserChannels is
an AuthRequest that requests only the user's enabled channels.

Make EPGFetcher.getChannels() return only the user's selected channels.
Rename the old getChannels() to getAllChannels().

Precursor to allowing batching of show requests to channel groups.

* Allow batching of EPG fetches to limit memory use

Allow the EPG to be fetched in batches of channels.

The aim is to reduce the maximum memory requirement when downloding
and processing the EPG. The larger (~140 channels, 14 days) German
EPG appeared to be causing OOM problems on some boxes.

config.plugins.icetv.batchsize controls the maximum size of the
batch.  When set to 0, the EPG is loaded in a single fetch as before.
Defaults to 30.

Bump version number.