[eEPGCache] Ensure that cache loads from file are non-overlapping
In the first loop of eEPGCache::load(), check whether events for a
channel are loaded in time order and are non-overlapping in time.

A test for non-overlapping that ignores time order would be more
complicated.

For any channel EPGs that may have time overlaps, loop over their
events and remove overlaps.

Ensuring that the EPG is initially non-overlapping is a prerequisite
for the following commit that significantly speeds up EPG event
import and EIT EPG event processing.