[eEPGCache] Speed up insertion of events into EPG
By taking advantage of the fact that the EPG per-channel list of
events by time is non-overlapping, the check for overlaps when a
new event is added in eEPGCache::sectionRead() can be reduced from
O(N) to O(log N), where N is the number of events in the channel's
EPG.

This improves both the processing speed of EIT EPG data and imported
EPG data (e.g. IceTV), because eEPGCache::sectionRead() is common
to both.

Also drop any zero-length events. Handling them would make
the speedup slightly more complicated.

For Australian IceTV data, it reduces the import time for the EPG
from ~0.67 sec to ~0.14 sec (4.8x), for an 8-day German IceTV EPG
(130+ channels), it reduces the import time from ~16.5 sec to ~2.32
sec (7.1x).

The production German IceTV data is expected to be 14 days, so the
improvement factor for that data should be greater.