Speed up insertion of events into EPG (#1453)
* [eEPGCache] Ensure that cache loads from file are non-overlapping

In the first loop of eEPGCache::load(), check whether events for a
channel are loaded in time order, non-zero length and are non-overlapping
in time.

A test for non-overlapping that ignores time order would be more
complicated.

For any channel EPGs that may have time overlaps or zero-, loop
over their events and remove overlaps.

Ensuring that the EPG is initially non-overlapping is a prerequisite
for the following commit that significantly speeds up EPG event
import and EIT EPG event processing.

Removing zero-length events slightly simplifies the code for the
speedup.

* [eEPGCache] Speed up insertion of events into EPG

By taking advantage of the fact that the EPG per-channel list of
events by time is non-overlapping, the check for overlaps when a
new event is added in eEPGCache::sectionRead() can be reduced from
O(N) to O(log N), where N is the number of events in the channel's
EPG.

This improves both the processing speed of EIT EPG data and imported
EPG data (e.g. IceTV), because eEPGCache::sectionRead() is common
to both.

Also drop any zero-length events. Handling them would make
the speedup slightly more complicated.

For Australian IceTV data, it reduces the import time for the EPG
from ~0.67 sec to ~0.14 sec (4.8x), for an 8-day German IceTV EPG
(130+ channels), it reduces the import time from ~16.5 sec to ~2.32
sec (7.1x).

The production German IceTV data is expected to be 14 days, so the
improvement factor for that data should be greater.