FileCommander: fix MessageBox on uneditable lines
Editing a line in File Commander's viewer would place the cursor at the
end of the line.  This is achieved by decoding as UTF-8 to get the
length.  However, if the decode fails the MessageBox warning that the
line is not editable would also fail due to modal issues.  Check the
line decodes before opening the edit dialog.