[Timezones.py] Patch to avoid /etc/localtime issues (#1568)
This patch disables the code that tried to synchronise the Enigma2 time zone with that defined in the Operating System.  For now the difference will be logged but Enigma2 will no longer be synchronized.

I hope this patch can be removed as soon as the code that is unexpectedly changing /etc/localtime can be found and fixed.