Implement enhancement #736: Improve checks for "._" files
The tests to ignore MacOS attribute files that start with "._" are done
by a test on the name returned by serviceHandler.info().getName().

This is incorrect, because the name returned by that process is often
a logical name divorced from the actual filename being tested.

The test should use the filename part of serviceref.getPath().

The changes implement that, and also change the position of the
test in the main directory scan loop of MovieList.load(), so that
if the filename starts with "._", serviceHandler.info() is never called
(since its returned value won't be used.

The changes are also a prerequisite for later changes that speed up
MovieList.dircount().

Also, change the test from name[:2] == "._" to name.startswith("._"_
for clarity.