Implement enhancement #736: Speed up MovieSelection directory counts
Calling self.getServiceInfo() on Beyonwiz recordings requires the
recording's .meta file to be read.

In MovieList.count() avoid calling self.getServiceInfo() unless necessary:
when tags are used and the recording's ligical name is needed, and to
get the name of a subdirectory.

This removes a lot of the overhead in count(), except if the user is
filtering using tags (when the cost is not avoidable).

A direct test on serviceref.getPath() is not done for mustDescent
servicerefs, because using the ServiceInformation path has the almost
the same overheads for servicerefs with type idFile, and it is required
for other serviceref types.