Update remote control management (#1592)
* [UsageConfig.py] Adjustments for refactored RcModel.py

This change removes the old "rc_model" settings.  These assume and constrain Enigma2 to use a configured remore control rather than using the most appropriate remote control as detected.

* [setup.xml] Remove "rc_model" option

This option is no longer used as the remote controls are determined at start  up and not from a previous selection.