Renderer patches (#1614)
* [Converter/ConditionalShowHide] Allows ConditionalShowHide to work in renderers that currently don't support it

e.g. render=Label".

This means we can now super-impose a text over a background colour and have the colour disappear if no text is present in the text field.

<widget source="key_yellow" render="Label" objectTypes="key_yellow,StaticText" position="710,686" size="240,25" zPosition="1" font="Regular;20" halign="left" transparent="0" valign="center" noWrap="1" foregroundColor="menu-buttons-frgrnd" backgroundColor="green">
 <convert type="ConditionalShowHide"/>
</widget>

Before this commit ConditionalShowHide didn't work. If screen was drawn with no text in the field the screen failed to display any widgets after that point.

Author: @prl001

* [Renderer.Label] Sanity test

This allows Label() to be used with non-text Sources

Fixes a bug with Renderer.Label when it tries to access source.text
when the upstream has no text attribute.

For example, in ViX-Common skin (used by ViX-Night-HD and ViX-Day-HD
skins) widget if you use this:

<widget source="VKeyIcon" conditional="VKeyIcon" render="Label" text="TEXT" position="1200,686" zPosition="1" size="35,25" transparent="1" alphatest="on">
 <convert type="ConditionalShowHide"/>
</widget>

Instead of the corresponding widget with a Pixmap renderer in
FullscreenConfigPanel, the TEXT label never displays and an exception
is thrown in the skin code "Error in screen 'Setup': 'ConditionalShowHide'
object has no attribute 'text'".

This changes Label.changed() to only try to access self.source.text if
the text attribute exists.

Author: @prl001

Co-authored-by: prl <prl@ozemail.com.au>